V03 Improved wrong date/time detection
Previous versions reported now and then false date/times. Simplified the valid time check and only reports a valid date/time if time does not differ more than 1 second and no more than one day from the previous received time. If one of the received pulses in a minute is <50 or >320 msec the time is also not reported